requirements.txt
torch
torchvision
torchaudio
rdkit
selfies
telebot
pubchempy
python-dotenv
pillow
scipy
tqdm
ipython_genutils
streamlit
py3Dmol
stmol
pandas
plotly
matplotlib
fpdf